feat: reorganize percentage calculation functionality and update documentation
