+++ dinero/tools/percentage.py
